requirements.txt
mysql-connector-python==9.0.0
customtkinter==5.2.2
python-dotenv==1.0.1
pillow==10.4.0
pywinstyles==1.8